New run handling
Much of the run and task handling which was previously dealt with by
lsc_tasker has been moved out into pycfd, so that this code may be used
to:
a)	Run code without enqueueing tasks on the tasking server
b)	Create tasks for running either LSC-AMR code or my pysolver code

In settings definitions have been moved into the modules for the two
solvers, as have task definitions for the two models. The settings
definitions also now contain the routines to store and retrieve setting
files.